Split provider.h in multiple include files
The file was getting long and difficult to manage.
